Listing name with uuid as helper
