+++ packages/catalog-realm-dev/catalog-app/listing/listing.gts
@@ -54,125 +54,6 @@ interface CopyMeta {
   targetCodeRef: ResolvedCodeRef;
 }
 
-function nameWithUuid(listingName?: string) {
-  if (!listingName) {
-    return '';
-  }
-  // sanitize the listing name, eg: Blog App -> blog-app
-  const sanitizedListingName = deburr(listingName.toLocaleLowerCase())
-    .replace(/ /g, '-')
-    .replace(/'/g, '');
-  const newPackageName = `${sanitizedListingName}-${uuidv4()}`;
-  return newPackageName;
-}
-
-export async function installListing(args: any, realm: string) {
-  const listing = args.model as Partial<Listing>;
-  const copyMeta: CopyMeta[] = [];
-
-  const localDir = nameWithUuid(listing.name);
-
-  // first spec as the selected code ref with new url
-  // if there are examples, take the first example's code ref
-  let selectedCodeRef: ResolvedCodeRef | undefined;
-  let shouldPersistPlaygroundSelection = false;
-  let firstExampleCardId;
-
-  // Copy the gts file based on the attached spec's moduleHref
-  for (const spec of listing.specs ?? []) {
-    const absoluteModulePath = spec.moduleHref;
-    const relativeModulePath = spec.ref.module;
-    const normalizedPath = relativeModulePath.split('/').slice(2).join('/');
-    const newPath = localDir.concat('/').concat(normalizedPath);
-    const fileTargetUrl = new URL(newPath, realm).href;
-    const targetFilePath = fileTargetUrl.concat('.gts');
-
-    await args.context?.actions?.copySource?.(
-      absoluteModulePath,
-      targetFilePath,
-    );
-    copyMeta.push({
-      sourceCodeRef: {
-        module: absoluteModulePath,
-        name: spec.ref.name,
-      },
-      targetCodeRef: {
-        module: fileTargetUrl,
-        name: spec.ref.name,
-      },
-    });
-
-    if (!selectedCodeRef) {
-      selectedCodeRef = {
-        module: fileTargetUrl,
-        name: spec.ref.name,
-      };
-    }
-  }
-
-  if (listing.examples) {
-    // Create serialized objects for each example with modified adoptsFrom
-    const results = listing.examples.map((instance: CardDef) => {
-      let adoptsFrom = instance[meta]?.adoptsFrom;
-      if (!adoptsFrom) {
-        return null;
-      }
-      let exampleCodeRef = instance.id
-        ? codeRefWithAbsoluteURL(adoptsFrom, new URL(instance.id))
-        : adoptsFrom;
-      if (!isResolvedCodeRef(exampleCodeRef)) {
-        throw new Error('exampleCodeRef is NOT resolved');
-      }
-      let maybeCopyMeta = copyMeta.find(
-        (meta) =>
-          meta.sourceCodeRef.module ===
-            (exampleCodeRef as ResolvedCodeRef).module &&
-          meta.sourceCodeRef.name === (exampleCodeRef as ResolvedCodeRef).name,
-      );
-      if (maybeCopyMeta) {
-        if (!shouldPersistPlaygroundSelection) {
-          selectedCodeRef = maybeCopyMeta.targetCodeRef;
-          shouldPersistPlaygroundSelection = true;
-        }
-
-        return {
-          sourceCard: instance,
-          codeRef: maybeCopyMeta.targetCodeRef,
-        };
-      }
-      return null;
-    });
-    const copyCardsWithCodeRef = results.filter(
-      (result) => result !== null,
-    ) as CopyCardsWithCodeRef[];
-    for (const cardWithNewCodeRef of copyCardsWithCodeRef) {
-      const { newCardId } = await args.context?.actions?.copyCard(
-        cardWithNewCodeRef.sourceCard,
-        realm,
-        cardWithNewCodeRef.codeRef,
-        localDir,
-      );
-      if (!firstExampleCardId) {
-        firstExampleCardId = newCardId;
-      }
-    }
-  }
-
-  if (listing instanceof SkillListing) {
-    await Promise.all(
-      listing.skills.map((skill) => {
-        args.context?.actions?.copyCard?.(skill, realm, undefined, localDir);
-      }),
-    );
-  }
-
-  return {
-    selectedCodeRef,
-    shouldPersistPlaygroundSelection,
-    firstExampleCardId,
-  };
-}
-
 export function setupAllRealmsInfo(args: any) {
   let allRealmsInfo =
     (args.context?.actions as Actions)?.allRealmsInfo?.() ?? {};
